requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
psycopg2-binary
python-dotenv
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1

